product_service/requirements.txt
flask==3.0.2
pandas==2.2.1
requests==2.31.0
gunicorn==21.2.0
langchain==0.1.16
langchain-community>=0.0.10
sentence-transformers==2.5.0
faiss-cpu==1.8.0
huggingface-hub==0.20.3
transformers==4.36.2
torch==2.1.2
cachetools==5.3.2
python-dotenv==1.0.0
pytest==7.4.3
